Fixed a long outstanding issue in gpu_prop_grid calling prop_grid_os2_trans_share_do. MVM now works in GPU as well as in CPU.

+++ cuda/pcg.h
@@ -20,7 +20,7 @@
 typedef void (*G_CGFUN)(curcell**, float, const void*, const curcell*, float);
 typedef void (*G_PREFUN)(curcell**, const void*, const curcell*, cudaStream_t stream);
 
-int gpu_pcg(curcell **px, 
+float gpu_pcg(curcell **px, 
 	    G_CGFUN Amul, const void *A, 
 	    G_PREFUN Mmul, const void *M, 
 	    const curcell *b, int warm, int maxiter,
